refactor(challenges): rename dropdown component to dropdown-filters
