+++ frontend/src/app/features/challenges/components/dropdown-filters/dropdown-filters.css
@@ -9,16 +9,16 @@
   justify-content: space-between;
   gap: 0.75rem;
   padding: 0.75rem 1.25rem;
-  border: 1px solid #d0d5dd;
+  border: 1px solid var(--color-gray-4);
   border-radius: 1.25rem;
-  background: white;
+  background: var(--color-white);
   cursor: pointer;
   position: relative;
   z-index: 1002;
 
   .button_text {
-    font-size: 1.5rem;
-    font-weight: 500;
+    font-size: var(--font-size-normal);
+    font-weight: var(--font-weight-regular);
   }
 
   .button_icon {
@@ -29,16 +29,16 @@
   .button_arrow {
     display: inline-block;
     transform: rotate(90deg);
-    font-size: 2em;
-    font-weight: 500;
-    line-height: 1;
+    font-size: var(--font-size-h2);
+    font-weight: var(--font-weight-regular);
+    line-height: var(--line-height-heading);
   }
 }
 
 .dropdown__overlay {
   position: fixed;
   inset: 0;
-  background: rgba(0, 0, 0, 0.5);
+  background: var(--color-transparent-black);
   z-index: 1001;
 }
 
@@ -51,7 +51,7 @@
   position: absolute;
   top: calc(100% + 1.5rem);
   right: 0;
-  background: white;
+  background: var(--color-white);
   border-radius: 0.5rem;
   padding: 1.5rem;
 }
