+++ frontend/src/app/features/challenges/components/dropdown-filters/dropdown-filters.html
@@ -1,14 +1,14 @@
 <div class="dropdown">
-  <button class="dropdown__trigger" type="button" (click)="toggleDropdown()">
+  <button class="dropdown__trigger" data-testid="dropdown-trigger" type="button" (click)="toggleDropdown()">
     <img src="assets\icons\filter-icon.png" class="button_icon" />
-    <span class="button_text">Filtros</span>
+    <span class="button_text">Filtres</span>
     <span class="button_arrow">›</span>
   </button>
 
   @if (isOpen()) {
-    <div class="dropdown__overlay" (click)="closeDropdown()"></div>
+    <div class="dropdown__overlay" data-testid="dropdown-overlay" (click)="closeDropdown()" ></div>
 
-    <div class="dropdown__menu">
+    <div class="dropdown__menu" data-testid="dropdown-menu">
       <button type="button">Fácil</button>
       <button type="button">Media</button>
       <button type="button">Difícil</button>
