+++ frontend/src/app/features/challenges/components/dropdown-filters/dropdown-filters.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { DropdownComponent } from './dropdown';
+import { DropdownComponent } from './dropdown-filters';
 
 describe('DropdownComponent', () => {
   let component: DropdownComponent;
@@ -23,30 +23,30 @@ describe('DropdownComponent', () => {
   it('should be closed by default', () => {
     expect(component.isOpen()).toBeFalsy();
 
-    const menu = fixture.debugElement.query(By.css('.dropdown__menu'));
-    const overlay = fixture.debugElement.query(By.css('.dropdown__overlay'));
+    const menu = fixture.debugElement.query(By.css('[data-testid="dropdown-menu"]'));
+    const overlay = fixture.debugElement.query(By.css('[data-testid="dropdown-overlay"]'));
 
     expect(menu).toBeNull();
     expect(overlay).toBeNull();
   });
 
   it('should open dropdown when trigger is clicked', () => {
-    const button = fixture.debugElement.query(By.css('.dropdown__trigger'));
+    const button = fixture.debugElement.query(By.css('[data-testid="dropdown-trigger"]'));
 
     button.nativeElement.click();
     fixture.detectChanges();
 
     expect(component.isOpen()).toBeTruthy();
 
-    const menu = fixture.debugElement.query(By.css('.dropdown__menu'));
-    const overlay = fixture.debugElement.query(By.css('.dropdown__overlay'));
+    const menu = fixture.debugElement.query(By.css('[data-testid="dropdown-menu"]'));
+    const overlay = fixture.debugElement.query(By.css('[data-testid="dropdown-overlay"]'));
 
     expect(menu).not.toBeNull();
     expect(overlay).not.toBeNull();
   });
 
   it('should close dropdown when trigger is clicked twice', () => {
-    const button = fixture.debugElement.query(By.css('.dropdown__trigger'));
+    const button = fixture.debugElement.query(By.css('[data-testid="dropdown-trigger"]'));
 
     button.nativeElement.click();
     fixture.detectChanges();
@@ -56,23 +56,23 @@ describe('DropdownComponent', () => {
 
     expect(component.isOpen()).toBeFalsy();
 
-    const menu = fixture.debugElement.query(By.css('.dropdown__menu'));
+    const menu = fixture.debugElement.query(By.css('[data-testid="dropdown-menu"]'));
     expect(menu).toBeNull();
   });
 
   it('should close dropdown when overlay is clicked', () => {
-    const button = fixture.debugElement.query(By.css('.dropdown__trigger'));
+    const button = fixture.debugElement.query(By.css('[data-testid="dropdown-trigger"]'));
 
     button.nativeElement.click();
     fixture.detectChanges();
 
-    const overlay = fixture.debugElement.query(By.css('.dropdown__overlay'));
+    const overlay = fixture.debugElement.query(By.css('[data-testid="dropdown-overlay"]'));
     overlay.nativeElement.click();
     fixture.detectChanges();
 
     expect(component.isOpen()).toBeFalsy();
 
-    const menu = fixture.debugElement.query(By.css('.dropdown__menu'));
+    const menu = fixture.debugElement.query(By.css('[data-testid="dropdown-menu"]'));
     expect(menu).toBeNull();
   });
 });
